fix: Changes comparing logic for `isAuthenticated` method
Access and ID tokens can have different lifetime. iOS OktaOidc library examines the expiration of these tokens. If the token is expired, null is returned.
Based on that we cannot return isAuthenticated=true because some of the tokens can be expired and refresh is needed. Otherwise, introspect throws an error evaluating null token.

Resolves: OKTA-367974 (#75)