fix: Changes comparing logic for `isAuthenticated` method
Access and ID tokens can have different lifetime. iOS OktaOidc library examines the expiration of these tokens. If the token is expired, null is returned.
Based on that we cannot return isAuthenticated=true because some of the tokens can be expired and refresh is needed. Otherwise, introspect throws an error evaluating null token.

Resolves: OKTA-367974 (#75)

OKTA-367974
<<<Jenkins Check-In of Tested SHA: 12d0656b41578b3787eca526404b9f7edc2112bc for eng_productivity_ci_bot_okta@okta.com>>>
Artifact: okta-react-native